Consolidate post-analysis artifact tests
